apps/api/requirements.txt
uvicorn
fastapi
websockets
google-genai
google-cloud-storage
httpx
python-dotenv
